requirements.txt
pandas>=1.5.0
scikit-learn>=1.2.0
joblib>=1.2.0
PyYAML>=6.0
